pyproject.toml
[tool.poetry]
name = "seo-keywork-planner"
version = "0.1.0"
description = ""
authors = ["Cauã Rinaldi <cauarinaldi451@gmail.com>"]
readme = "README.md"
packages = [{include = "seo_keywork_planner"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
python-dotenv = "^1.0.1"
google-auth-oauthlib = "^1.2.0"
google-ads = "^23.0.0"
pydantic = "^2.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
